# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/bridge/create": {
            "post": {
                "description": "创建一个新的消息中转配置。`is_active`默认为true（如果未提供）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bridge"
                ],
                "summary": "创建中转配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Bridge Create Object",
                        "name": "bridge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CreateBridgeInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Bridge"
                        }
                    },
                    "400": {
                        "description": "无效的输入或验证错误",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/bridge/delete/{id}": {
            "delete": {
                "description": "根据提供的ID删除消息中转配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bridge"
                ],
                "summary": "删除中转配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Bridge ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/bridge/get/{id}": {
            "get": {
                "description": "根据提供的ID获取单个消息中转配置的详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bridge"
                ],
                "summary": "根据ID获取中转配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Bridge ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Bridge"
                        }
                    },
                    "404": {
                        "description": "中转配置未找到",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/bridge/list": {
            "get": {
                "description": "获取所有消息中转配置的列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bridge"
                ],
                "summary": "获取所有中转配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Bridge"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/bridge/update/{id}": {
            "put": {
                "description": "根据提供的ID更新现有的消息中转配置。仅更新请求中提供的字段。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bridge"
                ],
                "summary": "更新中转配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Bridge ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Bridge Update Object",
                        "name": "bridge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.UpdateBridgeInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Bridge"
                        }
                    },
                    "400": {
                        "description": "无效的输入或验证错误",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.CreateBridgeInput": {
            "type": "object",
            "required": [
                "name",
                "source_channel_type",
                "target_channel_id",
                "ticket"
            ],
            "properties": {
                "is_active": {
                    "description": "使用指针以区分未提供和提供false的情况，默认为true",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "source_channel_type": {
                    "$ref": "#/definitions/types.ChannelType"
                },
                "target_channel_id": {
                    "type": "integer"
                },
                "ticket": {
                    "type": "string"
                }
            }
        },
        "controller.UpdateBridgeInput": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "source_channel_type": {
                    "$ref": "#/definitions/types.ChannelType"
                },
                "target_channel_id": {
                    "type": "integer"
                },
                "ticket": {
                    "type": "string"
                }
            }
        },
        "models.Bridge": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "source_channel_type": {
                    "$ref": "#/definitions/types.ChannelType"
                },
                "target_channel": {
                    "$ref": "#/definitions/models.Channel"
                },
                "target_channel_id": {
                    "type": "integer"
                },
                "ticket": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Channel": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "渠道配置（统一结构）",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "description": "UUID",
                    "type": "integer"
                },
                "name": {
                    "description": "渠道名称（如：Email-SMTP、OneBot-v11）",
                    "type": "string"
                },
                "type": {
                    "description": "渠道类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ChannelType"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "types.ChannelType": {
            "type": "string",
            "enum": [
                "Bark",
                "DingTalk",
                "FeiShu",
                "OneBot",
                "Gotify",
                "ServerChan",
                "PushDeer",
                "Synology",
                "IYUU",
                "PushPlus",
                "QMSG",
                "WeCom",
                "Telegram",
                "Email",
                "Webhook",
                "Ntfy",
                "WxPusher"
            ],
            "x-enum-comments": {
                "TypeBark": "Bark",
                "TypeDingTalk": "钉钉机器人",
                "TypeEmail": "邮箱",
                "TypeFeiShu": "飞书机器人",
                "TypeGotify": "Gotify",
                "TypeIYUU": "IYUU",
                "TypeNtfy": "Ntfy",
                "TypeOneBot": "OneBot",
                "TypePushDeer": "PushDeer",
                "TypePushPlus": "PushPlus",
                "TypeQMSG": "Qmsg酱",
                "TypeServerChan": "Server酱",
                "TypeSynology": "群晖chat",
                "TypeTelegram": "Telegram",
                "TypeWeCom": "企业微信",
                "TypeWebhook": "Webhook",
                "TypeWxPusher": "WxPusher"
            },
            "x-enum-varnames": [
                "TypeBark",
                "TypeDingTalk",
                "TypeFeiShu",
                "TypeOneBot",
                "TypeGotify",
                "TypeServerChan",
                "TypePushDeer",
                "TypeSynology",
                "TypeIYUU",
                "TypePushPlus",
                "TypeQMSG",
                "TypeWeCom",
                "TypeTelegram",
                "TypeEmail",
                "TypeWebhook",
                "TypeNtfy",
                "TypeWxPusher"
            ]
        },
...
    }
}